Finish up some job listing stuff
in createJobListing, we need to add "beingEdited" {boolean}, "isPublished" {boolean}, and "completed" {string options: "create", "step1", "step2", "step3", "step4", "step5", "appLimit", "payment"} fields to that mutation.

Add the fields "city" and "state" to the jobListing {under the main structure, not the business details} in the mutation step2.

Make a final mutation for publishing the listing and updating the isPublished field - it should return the job's id 🙂

Have those fields mentioned above returned as part of the jobListing, along with "applicationLimit" and "applications"

Remove the return of  the beingEdited mutation - we'll just need a simple response {like true/false} instead of an object 👍